refactor: normalize album photo identifier

+++ EduOS.Core/Entities/System/AlbumPhoto.cs
@@ -4,7 +4,7 @@ namespace EduOS.Core.Entities.System
 {
     public class AlbumPhoto : BaseEntity
     {
-        public int AlbumId { get; set; }
+        public long AlbumId { get; set; }
         public string ImageUrl { get; set; } = string.Empty;
         public string? Caption { get; set; }
         public DateTime UploadedAt { get; set; }
